requirements.txt
streamlit==1.31.1
openai==1.35.10
streamlit-extras==0.4.0
python-dotenv 
plotly
scipy
scikit-learn
pyodbc==4.0.35
SQLAlchemy==2.0.20
kaleido
prophet==1.1.5
statsmodels==0.14.1
tabulate==0.9.0
azure-search-documents==11.4.0
PyYAML==6.0.1  
tabulate
httpx
redis==5.0.8
fastapi==0.112.0
pydantic==2.7.4
uvicorn==0.30.5
azure-cosmos==4.7.0
requests-html
lxml_html_clean